Renamed appimaged.go to main.go
