+++ cmd/appimaged/inotify.go
@@ -29,7 +29,7 @@ import (
 
 	"github.com/fsnotify/fsnotify"
 	"github.com/probonopd/go-appimage/internal/helpers"
-	"github.com/probonopd/go-appimage/src/goappimage"
+	"github.com/probonopd/go-appimage/pkg/goappimage"
 )
 
 var watcher *fsnotify.Watcher
