+++ cmd/appimaged/main.go
